pyproject.toml
[project]
name = "autogentest1"
version = "0.1.0"
description = "Multi-agent AutoGen workflow for gold trading analytics powered by DeepSeek."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyautogen>=0.2.1",
    "pandas>=2.0",
    "numpy>=1.26",
    "yfinance>=0.2",
    "ta>=0.11",
    "feedparser>=6.0",
    "pydantic>=2.5",
    "python-dotenv>=1.0",
    "matplotlib>=3.8",
    "plotly>=5.18",
    "json-repair>=0.2",
    "requests-cache>=1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
